requirements.txt
fabric3
PyYAML
jinja2
